sit/requirements_311.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
Farama-Notifications==0.0.4
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.9.0
gast==0.6.0
glcontext==3.0.0
glfw==1.12.0
google-pasta==0.2.0
grpcio==1.75.1
gym==0.26.2
gym-notices==0.1.0
gym3==0.3.3
gymnasium==1.2.1
h5py==3.15.0
higher==0.2.1
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.3.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
kornia==0.8.1
kornia_rs==0.1.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
moderngl==5.12.0
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
procgen2==0.10.7
protobuf==6.33.0
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.2.0
scipy==1.16.2
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
triton==3.5.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
